r1610 | dbaxter | 2012-08-05 11:56:33 -0700 (Sun, 05 Aug 2012) | 11 lines
Initial concentrations file format changed to accommodate compartments,
and the notion of fixed or variable concentration for a molecule. Format
for lines in the initial concentrations file is
molecule:compartment concentration [V|C]

where [V|C] means one of those characters with out the brackets.
the :compartment is optional and is the V or C. If V or C is not
specified V is assumed.

M    read_initial_concentrations.c