r3723 | dbaxter | 2016-12-12 16:11:18 -0800 (Mon, 12 Dec 2016) | 7 lines
This used to be approximate_delta_concs, and before that
approximate_fluxes. It approxmiates the concentration change with
respect to time using likelihood ratio approximations and under the
assumption that the reverse reaction likelihood ratios are all the
same as the base reactions reverse reaction likelihood ratio.
It is the default approximate_delta_concs evaluater.