r4422 | dbaxter | 2017-07-11 07:45:11 -0700 (Tue, 11 Jul 2017) | 12 lines
Added counts_or_concs and count_or_conc arguments,
counts_or_concs is a vector of counts if count_or_conc = 1, of
concentrations if count_or_conc = 0.
Changed variable names to be more informative,
and now raise the regulation constant to the regulation power as well as
the concentration.
Also modified the compartment to come from the compartment size of the
regualtion species. Before it was based on the compartment of the species
of the reaction number which was just wrong, but probably never mattered in
single compartment calculations.
M    update_regulation.c