r4437 | dbaxter | 2017-08-15 15:52:05 -0700 (Tue, 15 Aug 2017) | 11 lines
New concentration derivative formulation based on concentrations and
equilibrium coefficients.

for reaction j if species i is a product in the reaction.
 d[species_i]/dt =  1/gamma_i * (k_{j} * P_r  - k_{-j} * P_p)

P_r and P_p are products reactant concentrations and product concentrations
respectively and k_{j} and k_{-j} are the equilibrium constants of the
forward and reverse reactions.
M    lr9_approximate_delta_concs.c